+++ src/main/java/com/hnptech/stocknewscuckoo/article/dto/response/ArticleResponse.java
@@ -1,8 +1,5 @@
-package com.hnptech.stocknewscuckoo.article.model;
+package com.hnptech.stocknewscuckoo.article.dto.response;
 
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
 import java.time.LocalDateTime;
 import lombok.AccessLevel;
 import lombok.Builder;
@@ -11,21 +8,16 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Entity
-public class Article {
+public class ArticleResponse {
 
-	@Id
 	private String url;
-
 	private String title;
-
-	// 미국 시간 기준
 	private LocalDateTime publishedAt;
 
 	@Builder
-	public Article(String title, String url, LocalDateTime publishedAt) {
-		this.title = title;
+	public ArticleResponse(String url, String title, LocalDateTime publishedAt) {
 		this.url = url;
+		this.title = title;
 		this.publishedAt = publishedAt;
 	}
 }
