+++ src/main/java/com/hnptech/stocknewscuckoo/article/mapper/ArticleMapper.java
@@ -1,0 +1,11 @@
+package com.hnptech.stocknewscuckoo.article.mapper;
+
+import com.hnptech.stocknewscuckoo.article.dto.response.ArticleResponse;
+import com.hnptech.stocknewscuckoo.article.model.Article;
+import java.util.List;
+
+public interface ArticleMapper {
+	ArticleResponse toResponse(Article article);
+	List<ArticleResponse> toResponseList(List<Article> articles);
+
+}
