+++ src/main/java/com/hnptech/stocknewscuckoo/article/service/ArticleService.java
@@ -1,11 +1,12 @@
 package com.hnptech.stocknewscuckoo.article.service;
 
+import com.hnptech.stocknewscuckoo.article.dto.response.ArticleResponse;
 import com.hnptech.stocknewscuckoo.article.model.Article;
 import java.util.List;
 
 public interface ArticleService {
 
-	List<Article> getLatestArticles();
+	List<ArticleResponse> getLatestArticles();
 
 	void saveArticles(List<Article> articles);
 }
