+++ src/main/java/com/hnptech/stocknewscuckoo/article/service/ArticleServiceImpl.java
@@ -1,5 +1,7 @@
 package com.hnptech.stocknewscuckoo.article.service;
 
+import com.hnptech.stocknewscuckoo.article.dto.response.ArticleResponse;
+import com.hnptech.stocknewscuckoo.article.mapper.ArticleMapper;
 import com.hnptech.stocknewscuckoo.article.model.Article;
 import com.hnptech.stocknewscuckoo.article.repository.ArticleRepository;
 import java.util.List;
@@ -11,14 +13,12 @@ import org.springframework.stereotype.Service;
 public class ArticleServiceImpl implements ArticleService {
 
 	private final ArticleRepository articleRepository;
-
+	private final ArticleMapper articleMapper;
 	@Override
-	public List<Article> getLatestArticles() {
-		return articleRepository.findTop20ByOrderByPublishedAtDesc();
+	public List<ArticleResponse> getLatestArticles() {
+		 return articleMapper.toResponseList(articleRepository.findTop20ByOrderByPublishedAtDesc());
 	}
 
-
-	//TODO : 중복된 기사가 저장됨 수정해야함
 	@Override
 	public void saveArticles(List<Article> articles) {
 		List<Article> notDuplicatedArticles = articles.stream()
@@ -27,4 +27,6 @@ public class ArticleServiceImpl implements ArticleService {
 
 		articleRepository.saveAll(notDuplicatedArticles);
 	}
+
+	//TODO : 중복 기사 데이터 확인 최적화, 상수 관리, Get 쿼리 최적화
 }
